Move donut to main ui directory
